Floating Away

+++ edit_version.py
@@ -5,7 +5,7 @@ with open("version.txt", "r") as r:
     print(f"* {currentversion} is the current version")
     r.close()
 with open("version.txt", "w") as w:
-    latestversion = str(int(currentversion) + 0.01)
+    latestversion = str(float(currentversion) + 0.01)
     print(f"* changed version to {latestversion}")
     w.write(latestversion)
     w.close()
